Added support for non-primitive-value-referencing queries.
Query methods can now query for related objects of an entity. Test cases are centered around a Person having an Address. The query is something like:

List<Person> findByAddress(Address address);

We correctly marshal the parameter into a DBObject using the MongoConverter on argument binding.

Exposed the configured MongoConverter at MongoTemplate so that the repository infrastructure gets access to it. Let repository classes work with the template again instead of plain MongoOperations. Removed find(…) method from MongoOperations and -Template. Extended SimpleMongoConverter so that it can unmarshal collections as well.