DATAMONGO-725 - Improve configurability and documentation of TypeMapper on MappingMongoConverter.
Removed the unsupported attributes "mongo-ref" and "mongo-template-ref" from the mapping-converter element in spring-mongo-1.3.xsd because they are not considered anymore.
Added new attribute type-mapper-ref to the mapping-converter element in spring-mongo-1.3.xsd in order to support the configuration of custom-type-mappers.
Updated MappingMongoConverterParser to be aware of the new attribute.

Added examples for configuring a custom MongoTypeMapper the usage of @TypeAlias to the reference documentation.