DATAMONGO-774 - General jürgenization of the expression transformation code.
Introduced dedicated spel package and extracted value objects to encapsulate and express information about the node transformation in a more semantical way. 

Moved a lot of the logic contained in the SpelExpressionTransformer into the value objects for cohesiveness and testability.

Updated Sonargraph architecture model to reflect the new packages we've introduced.