DATAMONGO-807 - findAndModify does not retain Class information.
Using findAndUpdate did not retain _class information when used to update a whole nested type instead of single fields within the type. We now use the updateMapper in doFindAndModify for the update operation to make sure to retain the type information - previously we just used the queryMapper.

Original pull request: #110