DATAMONGO-407 - Collection with generics losing element type after $set update.
Introduced UpdateMetadataBackedField in UpdateMapper to be able to properly deal with collection element updates. In this case we have to maintain the original field path that contains potential collection element wildcard parts as well as a stripped / linearised version of the property path that doesn't contain element wildcard parts at all to be able to correctly determine the proper Type of the property.
We now use the UpdateMapper in MongoTemplate#doChangeAndModify(..) to write the correct type information for Update objects.