DATAMONGO-843 - Make sure to reference the correct MongoMappingContext internally.
We now use an ObjectFactory<MappingContext> as the constructor argument in order to avoid a dependency cycle. This cycle arises from the initialisation of the MappingContext (afterPropertiesSet()) triggers the bean creation of the AuditingEventListener due to the eager initialisation of ApplicationEventListeners by the spring framework. This implicitly requests the bean creation of the MappingContextIsNewStrategy which requires the initial MappingContext that is still in construction which forms the cycle.

This depends on the changes in DATACMNS-455 to be applied.