DATAMONGO-884 Check for Paramter obj is null
It seems when intercept function is called the parameter "Object obj" is
always "null" (in my case), which result NPE when Object method is
called.

If the parameter Object obj is actually always null, it  either should
remove all the checks and just always invoke by
"method.invoke(ensureResolved(), args);"

If there is case that parameter Object obj is NOT null, then it should
add another check to handle the case, which the parameter Object obj is
null.