DATAMONGO-895 - Use most specific type for checks against values in DBObjects.
When comparing the value of the version field, compare against a Long rather than an Integer, since the version field generated is a Long. This allows the test to pass against the upcoming 2.12.0 release of the Java driver, which has a stricter implementation of BasisDBObject.equals.

Original Pull Request: #159 (Jeff Yemin)