DATAMONGO-897 - DBRef Update mapping should work on interface types.
Fixed potential failure in DBRef resolution for updates.

In 1.4 'DBRef' resolution in 'QueryMapper' changed in a way that tries to inspect if the properties 'Id' can be assigned to the given type. When one passes in an 'Update' it is most likely not to have all type information at hand, which means in case of an property pointing to an interface that the entire resolution might have failed.

In 1.5 it's possible to resolve properties on interfaces as long as those follow the beans convention. But it failed, as it does in 1.4.x, for those properties not having the required getters.

Original Pull Request: #164