DATAMONGO-888 - Mapping should be applied to Sort.
'EntityBackedSortConverter', as an instance of 'SortConverter', takes care of mapping properties to the according names to be used with mongodb.

Conversion will inspect properties within the backing entity to resolve ids and additional mapping information within the 'Field' annotation and apply those to the resulting 'DBObject'.

The converter will not touch any properties unknown to the backing entity.

'Query.getSortObject()' has been deprecated and will be removed in a subsequent version.

Original Pull Request: #162