DATAMONGO-914 - Improve resolving of LazyLoading proxies for classes that override equals/hashcode.
We now properly resolve lazy loading proxies for @DBRef's when an overridden equals or hash code method is called with Spring 4.
We fall back to our old Objenesis proxy generation in order to circumvent the default handling for overridden hashcode and equals methods in CglibAopProxies generated by Spring 4. If we detect that we run with Spring 4 we use the repacked Objenesis that is included in Spring 4.
Previously the generated proxy used some generic hashcode or equals logic that did not trigger a proper lazy loading in such cases.

Original pull request: #171.