DATAMONGO-847 - Allow usage of Query for Update $pull.
In case we detect `Query` within a value used for an `Update` value we map the query itself to build the expression to use. This allows to form query statements for eg. `$pull` using the same API as for the query itself. 

Update update = new Update().pull("list", query(where("value").in("foo", "bar")));

Original Pull Request: #172