DATAMONGO-917 - Improve Spring 4.0 framework version detection to avoid NPEs.
We now check for the presence of "DefaultParameterNameDiscoverer" in order to determine if we are running with a Spring version later than 4.0 since this avoids potential NPEs in cases where the package version information is not available e.g. in cases where the application was bundled into an "uberjar" e.g. via the maven-shade-plugin.

Original pull request: #173.